+++ MarkdownParser/Antlr/Answer/Blocks/ParagraphBlock.cs
@@ -1,27 +1,18 @@
 ﻿using MarkdownParser.Antlr.Answer.Parts;
 using System.Windows;
 using System.Windows.Documents;
-using System.Windows.Media;
 
-namespace MarkdownParser.Antlr.Answer
+namespace MarkdownParser.Antlr.Answer.Blocks
 {
-    public sealed class Block
+    public sealed class ParagraphBlock : IBlock, ITextualBlock
     {
-        private static readonly System.Windows.Media.Brush _semiTransparentGray = new SolidColorBrush(Color.FromArgb(0x40, 0x80, 0x80, 0x80));
-
         private readonly List<IPart> _parts = new();
         private readonly IFontSizeProvider _fontSizeProvider;
         private BlockUIContainer? _blockContainer;
 
-        public BlockTypeEnum Type
-        {
-            get;
-            private set;
-        }
+        public BlockTypeEnum Type => BlockTypeEnum.Paragraph;
 
-        public IReadOnlyList<IPart> Parts => _parts;
-
-        public Block(
+        public ParagraphBlock(
             IFontSizeProvider fontSizeProvider
             )
         {
@@ -31,19 +22,7 @@ namespace MarkdownParser.Antlr.Answer
             }
 
             _fontSizeProvider = fontSizeProvider;
-            
-            Type = BlockTypeEnum.Paragraph;
         }
-
-        public void SetType(
-            BlockTypeEnum type,
-            BlockUIContainer? blockContainer
-            )
-        {
-            Type = type;
-            _blockContainer = blockContainer;
-        }
-
 
         public System.Windows.Documents.Block CreateBlock(
             AdditionalCommandContainer? acc,
@@ -55,13 +34,12 @@ namespace MarkdownParser.Antlr.Answer
                 return _blockContainer;
             }
 
-            var paragraph = new System.Windows.Documents.Paragraph
+            var paragraph = new Paragraph
             {
                 Margin = new Thickness(10, 10, 0, 0),
             };
-            ModifyParagraph(paragraph);
 
-            foreach (var part in Parts)
+            foreach (var part in _parts)
             {
                 foreach (var inline in part.GetInlines(isInProgress))
                 {
@@ -76,31 +54,6 @@ namespace MarkdownParser.Antlr.Answer
             }
 
             return paragraph;
-        }
-
-        private void ModifyParagraph(Paragraph paragraph)
-        {
-            switch (Type)
-            {
-                case BlockTypeEnum.Paragraph:
-                    return;
-                case BlockTypeEnum.Blockquote:
-                    paragraph.Margin = new Thickness(
-                        paragraph.Margin.Left,
-                        0,
-                        paragraph.Margin.Right,
-                        paragraph.Margin.Bottom
-                        );
-                    paragraph.Background = _semiTransparentGray;
-                    paragraph.BorderBrush = Brushes.Green;
-                    paragraph.BorderThickness = new Thickness(5, 0, 0, 0);
-                    paragraph.Padding = new Thickness(5, 5, 5, 5);
-                    return;
-                case BlockTypeEnum.HorizontalRule:
-                    return;
-                default:
-                    throw new InvalidOperationException(Type.ToString());
-            }
         }
 
         public void AddText(string text)